+++ src/PS.FarNet.Redis/ExportCommand.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Management.Automation;
 
 namespace PS.FarNet.Redis;
@@ -14,6 +15,17 @@ public sealed class ExportCommand : BaseDBCmdlet
     public string Pattern { get; set; }
 
     [Parameter]
+    [SupportsWildcards]
+    public string[] Exclude
+    {
+        set
+        {
+            _excludePatterns = value.Select(x => new WildcardPattern(x)).ToArray();
+        }
+    }
+    WildcardPattern[] _excludePatterns;
+
+    [Parameter]
     public TimeSpan? TimeToLive { get; set; }
 
     protected override void BeginProcessing()
@@ -26,6 +38,7 @@ public sealed class ExportCommand : BaseDBCmdlet
         {
             Pattern = Pattern,
             TimeToLive = TimeToLive,
+            Exclude = _excludePatterns is null ? null : key => _excludePatterns.Any(x => x.IsMatch(key)),
             WriteWarning = WriteWarning,
         };
         command.Invoke();